cleanup of buff msgs, added direct pkt write methods for utf16 vectors open question: keep it this way or add an string write method which converts on the fly?

+++ pol-core/pol/mobile/charactr.h
@@ -208,7 +208,8 @@ struct Buff
   u32 cl_descr;
   // Unicode string, arguments to be replaced in cl_descr, separated by tabs
   // TODO: use a real unicode string class, maybe a vector of them
-  std::vector<u32> arguments;
+  // or not since here the correct format already exists
+  std::vector<u16> arguments;
 };
 
 struct reportable_t
@@ -706,7 +707,7 @@ protected:
   // BUFF/DEBUFF BAR
 public:
   void addBuff( u16 icon, u16 duration, u32 cl_name, u32 cl_descr,
-                const std::vector<u32>& arguments );
+                const std::vector<u16>& arguments );
   bool delBuff( u16 icon );
   void clearBuffs();
   void send_buffs();
